Add sbt-scoverage plugin
To run tests with coverage enabled and generate a report, execute:

    $ sbt clean coverage test coverageReport

If your tests failed for some reason, it is very likely that the report
generated by scoverage will be useless. You need to fix the tests and
re-run the above command.

Test reports in HTML and XML formats can be found in
/target/scala-2.12/scoverage-reports/ directory.

Signed-off-by: Nikolai Obedin <nikolai.obedin@here.com>