Fix launching retry-campaign when a device fails with a new code
Currently, if a device fails in a retry-campaign with a failure code
that is different than the one it originally failed with, we are not
able to take it into account when launching a new retry-campaign for
the new failure code. Instead we get a MissingFailedDevices error
because we're only searching for devices that failed under the main
campaign. This commit fixes that and adds a test for it.

Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>