Do not update the status of a campaign if it was cancelled
Previously the status of a campaign was updated each time a new device
finishes. A campaign could get `cancelled` status when it had all its
groups cancelled. When switching from using groups to sets of
devices, this functionality was lost and a campaign could never get
`cancelled` status based on statuses of its devices.

This PR disables updating campaign's status when it's set as `cancelled`
by the code in `/api/v2/campaigns/<uuid>/cancel` endpoint.

NOTE: the status won't be changed to `failed` even if some cancelled
devices fail. Once it's cancelled, it won't EVER be changed. If this
behaviour is not desirable, e.g. the campaign should get `failed` status
even though it was `cancelled` earlier, it could be fixed.

Signed-off-by: Nikolai Obedin <nikolai.obedin@here.com>