Device can cancel update coming from campaign pending approval
Campaigner should only offer an endpoint to cancel updates for a device
in a campaign that is pending approval, since these are not known to
director yet, i.e. there is no assignment in director.

For the other cases, when an update has been assigned in director (a
campaign update or a MTU) we should call director at DELETE
`/api/v1/assignments/:deviceUuid` to cancel the update. Director will
publish a `DeviceUpdateCancelled` event and, if the update came from a
campaign, campaigner process the event and cancel the update.

Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>