Cancel device updates that were never scheduled
This fixes a bug that exists when a user cancels a campaign before the
scheduler pushes the campaign to director. In that case the Cancel
task would run, but not pickup "requested" devices, since those would
not need to be canceled in director, but it would also skip setting
the cancel status on the corresponding device update, which means
Campaign Scheduler keeps picking it up forever on it's polling of the db.

Also drain the namespace config topic before starting the daemons, so
that the daemons use the proper director from the start.

Signed-off-by: Simão Mata <simao.mata@here.com>