Fix bug where campaigns keep being picked up by the supervisor.
This partially reverts 84ae6632. The strange looking query there was actually
keeping the campaign supervisor actor from picking up and scheduling campaigns
indefinitely. When I removed it the daemon went mental trying to schedule always
the same campaigns, and it was using more CPU and Slick threads than before.

So to counter that, I'm bringing back the query but filtering only on launched
campaigns. So the behaviour should be the same than before, while not picking up
and launching campaigns automatically, which is the bug I was trying to fix in
the first place.

In my opinion this is not the best solution, and a better approach could be to
introduce a new `inProgress` status for the campaign. The supervisor would find
all `launched` campaigns (as it does before this commit) and then set the campaign
to `inProgress` once it's scheduled. But that would require some arduous refactor
and probably some changes to the UI regarding the campaign status, so I'd rather
go with a minimal change.

Relates-to: OTA-4951
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>